+++ backend/setup.sh
@@ -1,24 +1,21 @@
 #!/bin/bash
 
-MVIS_BACKEND_ADDRESS = $1
-SCHEMA_PATH = $2
-
 function setup_schema {
-    curl --silent -X POST "${MVIS_BACKEND_ADDRESS}/admin/schema" --data-binary "@${SCHEMA_PATH}"
+    curl --silent -X POST "${DEPLOYMENT_URL}/admin/schema" --data-binary "@/app/schema.graphql"
     echo ""
 }
 
-until $(curl --output /dev/null --silent ${MVIS_BACKEND_ADDRESS}/health || exit 1); do
-    echo "${MVIS_BACKEND_ADDRESS} is not active"
+until $(curl --output /dev/null --silent ${DEPLOYMENT_URL}/health || exit 1); do
+    echo "${DEPLOYMENT_URL} is not active"
     sleep 5
 done
 
-LTIME=$(stat -c %Z ${SCHEMA_PATH})
+LTIME=$(stat -c %Z /app/schema.graphql)
 
 setup_schema
 
 while sleep 1; do
-    ATIME=$(stat -c %Z ${SCHEMA_PATH})
+    ATIME=$(stat -c %Z /app/schema.graphql)
 
     if [[ "$ATIME" != "$LTIME" ]]; then
         setup_schema
